requirements.txt
requests
beautifulsoup4
playwright
pillow
numpy
scikit-learn
cairosvg
pandas
tqdm
urllib3